todo: also Abc_Ntk_t

+++ src/aig/aig/aig.h
@@ -288,7 +288,7 @@ static inline int          Aig_ObjIsChoice( Aig_Man_t * p, Aig_Obj_t * pObj )   
 static inline int          Aig_ObjIsCand( Aig_Obj_t * pObj )      { return pObj->Type == AIG_OBJ_CI || pObj->Type == AIG_OBJ_AND || pObj->Type == AIG_OBJ_EXOR;     }
 static inline int          Aig_ObjCioId( Aig_Obj_t * pObj )       { assert( !Aig_ObjIsNode(pObj) ); return pObj->CioId;                                            }
 static inline int          Aig_ObjId( Aig_Obj_t * pObj )          { return pObj->Id;                     }
-static inline int          Aig_ObjCertifId( Aig_Obj_t * pObj )    { return pObj->CertifId;               } //@ getter
+static inline uint64_t     Aig_ObjCertifId( Aig_Obj_t * pObj )    { return pObj->CertifId;               } //@ getter
 
 static inline int          Aig_ObjIsMarkA( Aig_Obj_t * pObj )     { return pObj->fMarkA;  }
 static inline void         Aig_ObjSetMarkA( Aig_Obj_t * pObj )    { pObj->fMarkA = 1;     }
